Fix penalty time
The previous change for long penalty time was incomplete and causes the penalty to be 0 on the scoreboard. This fixes that problem and follows references to make sure that Result is also fully switched over to long ms time values. Standings and all other classes are left with penalty as an integer (minute) value.

+++ ContestModel/src/org/icpc/tools/contest/model/IResult.java
@@ -41,11 +41,11 @@ public interface IResult {
 	long getContestTime();
 
 	/**
-	 * Return the penalty time for this problem, in minutes.
+	 * Return the penalty time for this problem.
 	 *
-	 * @return the penalty time
+	 * @return the penalty time, in ms.
 	 */
-	int getPenaltyTime();
+	long getPenaltyTime();
 
 	/**
 	 * Return the score for this problem.
